Add LICENSE (MIT, following setup.py)
In order for python-pure-cdb to be packaged by a distribution like
Debian, it needs a clear license. Debian really needs python-pure-cdb
because the python-cdb package has been removed due to a lack of
licensing clarity there.

The setup.py file contains a `license='MIT'` line, from which I deduce
that the LICENSE file should contain the text of the MIT license. The
copyright date range I derived from the first and last commits.

It looks as if Daniel Néri <dne@mayonnaise.net> and Philippe
Ombredanne <pom@nexb.com> have contributed to this project. Getting an
Ack from them on this commit would be good.

Signed-off-by: Douglas Bagnall <douglas@halo.gen.nz>