Remove advice to use hash() instead of djb_hash()
The trouble is not only that the hash function is subject to change
between versions of Python, but also that the current (2.7+) version
of the hash function is not necessarily stable across processes:

    $ export PYTHONHASHSEED=random
    $ for x in 1 2 3; do python -c 'print(hash("a"))';done
     -6203624181990238437
     3653280207417954465
     -928541188491112573

Randomised hashing depends on the environment variable or the -R flag
in Python 2.7, and is on by default in Python 3.

The compiled version of djb_hash is probably not so bad anyway. I
haven't tested.